refactor: change usages to scripts.db.utils.*
